Revert "Appコンポーネントを分割"
