--- src/GlobalStyle.tsx
@@ -1,15 +1,0 @@
-import { createGlobalStyle } from 'styled-components'
-
-export const GlobalStyle = createGlobalStyle`
-  html, body, #app {
-    height: 100%;
-  }
-
-  body {
-    /* https://css-tricks.com/snippets/css/system-font-stack/ */
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto,
-      Oxygen-Sans, Ubuntu, Cantarell, 'Helvetica Neue', sans-serif;
-
-    overflow-wrap: break-word;
-  }
-`
